+++ samples/SampleWebApi/Controllers/Issue80.cs
@@ -1,4 +1,6 @@
-﻿using System.Text.Json.Serialization;
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using System.Text.Json.Serialization;
 using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 
@@ -31,6 +33,35 @@ namespace SampleWebApi.Controllers
                 RuleFor(bestShot => bestShot.Link)
                     .MinimumLength(5);
             }
+        }
+
+        //88
+
+        public class BulkSoftwareDeliveryCreateModel
+        {
+            [Required]
+            [MinLength(3)]
+            [MaxLength(100)]
+            public string Name { get; set; } = null!;
+
+            [MaxLength(255)]
+            public string Description { get; set; } = null!;
+
+            [Required]
+            [MinLength(3)]
+            [MaxLength(100)]
+            public string ProductDeliveryGroupCode { get; set; } = null!;
+
+            [Required]
+            [RegularExpression(@"^((([0-9]{1,4})\.([0-9]{1,4})\.([0-9]{1,4})))$")]
+            public string DisplayVersion { get; set; } = null!;
+
+            [Required]
+            public DateTime? ReleaseDate { get; set; } // Must be nullable/required to prevent a default value being used.
+
+            public string? Languages { get; set; }
+
+            public string? Platforms { get; set; }
         }
     }
 }